+++ lib/modules/home/screens/search.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:algolia/algolia.dart';
 import 'package:andax/models/story.dart';
 import 'package:andax/modules/home/services/searching.dart';
-import 'package:andax/modules/home/services/stories.dart';
 import 'package:andax/shared/extensions.dart';
 import 'package:andax/shared/widgets/paging_list.dart';
 import 'package:andax/shared/widgets/rounded_back_button.dart';
