+++ lib/modules/home/services/searching.dart
@@ -1,4 +1,5 @@
 import 'package:algolia/algolia.dart';
+import 'package:andax/models/story.dart';
 import 'package:andax/store.dart';
 
 String _generateFilter(
@@ -28,4 +29,20 @@ AlgoliaQuery formQuery(String index, String text) {
     query = query.filters(_generateFilter(tags, 'tags', true));
   }
   return query;
+}
+
+Future<List<StoryInfo>> getStories(
+  String index, {
+  int? page,
+  int? hitsPerPage,
+}) async {
+  var query = algolia.instance.index(index).query('');
+  if (page != null) query = query.setPage(page);
+  if (hitsPerPage != null) query = query.setHitsPerPage(hitsPerPage);
+  final qs = await query.getObjects();
+  return storiesFromSnapshot(qs);
+}
+
+List<StoryInfo> storiesFromSnapshot(AlgoliaQuerySnapshot qs) {
+  return qs.hits.map((h) => StoryInfo.fromAlgoliaHit(h)).toList();
 }
