+++ lib/modules/home/widgets/gradient_cover_image.dart
