+++ lib/modules/profile/screens/profile.dart
@@ -1,7 +1,7 @@
 import 'package:andax/models/story.dart';
+import 'package:andax/modules/home/services/sheets.dart';
 import 'package:andax/shared/widgets/paging_list.dart';
 import 'package:andax/shared/widgets/rounded_back_button.dart';
-import 'package:andax/shared/widgets/story_sheet.dart';
 import 'package:andax/shared/widgets/story_tile.dart';
 import 'package:andax/store.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
