+++ lib/modules/story_editor/widgets/actor_picker_sheet.dart
@@ -1,10 +1,11 @@
 import 'package:andax/models/actor.dart';
 import 'package:andax/modules/story_editor/screens/story_editor.dart';
 import 'package:andax/modules/story_editor/widgets/actor_tile.dart';
-import 'package:andax/shared/widgets/modal_scrollable_sheet.dart';
 import 'package:andax/shared/widgets/rounded_back_button.dart';
+import 'package:andax/shared/widgets/scrollable_modal_sheet.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+
 import 'actor_editor_dialog.dart';
 
 void showActorPickerSheet(
@@ -13,7 +14,7 @@ void showActorPickerSheet(
   String? selectedId,
 ]) {
   final editor = context.read<StoryEditorState>();
-  showModalScrollableSheet<Actor>(
+  showScrollableModalSheet<Actor>(
     context: context,
     builder: (context, scroll) {
       final actors = editor.story.actors.values.toList();
