+++ lib/modules/story_editor/widgets/node_picker_sheet.dart
@@ -2,8 +2,8 @@ import 'package:andax/models/node.dart';
 import 'package:andax/modules/story_editor/screens/node_editor.dart';
 import 'package:andax/modules/story_editor/screens/story_editor.dart';
 import 'package:andax/modules/story_editor/widgets/node_tile.dart';
-import 'package:andax/shared/widgets/modal_scrollable_sheet.dart';
 import 'package:andax/shared/widgets/rounded_back_button.dart';
+import 'package:andax/shared/widgets/scrollable_modal_sheet.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -13,7 +13,7 @@ Future<Node?> showNodePickerSheet(
   String? selectedId,
 ]) {
   final editor = context.read<StoryEditorState>();
-  return showModalScrollableSheet(
+  return showScrollableModalSheet(
     context: context,
     builder: (context, scroll) {
       final nodes = editor.story.nodes.values.toList();
